backend/requirements.txt
Flask==3.1.1
python-dotenv==1.0.1
Flask-SQLAlchemy
psycopg2-binary
flask_jwt_extended

